+++ HISTORY.md
@@ -1,4 +1,5 @@
 ## 0.0.16
+- Added a script "gandlf_deploy", allowing deployment of models into MLCubes (currently requires Docker)
 - ImageNet pre-trained models for UNet with variable encoders is now available
 - ACS/Soft conversion is available for ImageNet-pretrained UNet
 - Updated links, copyright and email to MLCommons
@@ -6,6 +7,7 @@
 - Added ability to combine classification inference results from different architectures
 - Added ability to save `initial` and `latest` models in addition to `best`
 - Added ability to specify testing data csv in main cli
+- Normalized surface dice has been added
 
 ## 0.0.15
 - Updated `setup.py` for `python>=3.8`